feat: add Pet component
